+++ README.md
@@ -17,7 +17,6 @@
 - 우승자 판별
 - 우승자 이름 출력
 - 자동차 이름을 잘못 입력할 경우 alert 메시지 띄우기
-- 횟수를 잘못 입력할 경우 alert 메시지 띄우기
 
 ## 🎯 기능 요구사항
